ULTIMA8: Fix parameter type and bitmask for AAF_ROTATED flag

+++ engines/ultima/ultima8/world/actors/anim_action.cpp
@@ -87,7 +87,7 @@ const AnimFrame &AnimAction::getFrame(Direction dir, unsigned int frameno) const
  having to check against game type each time they are used
 */
 /*static*/
-AnimAction::AnimActionFlags AnimAction::loadAnimActionFlags(uint16 rawflags) {
+AnimAction::AnimActionFlags AnimAction::loadAnimActionFlags(uint32 rawflags) {
 	uint32 ret = AAF_NONE;
 	ret |= (rawflags & AAF_COMMONFLAGS);
 	if (GAME_IS_U8) {
